Move storage codes to storage folder
